Fix compile time error
It's going to throw you a compile time error if you try assigning it to a string. I suspect that this hasn't been updated in a while where as the actual library has. I'm also letting the user know how to generate their own API key by referencing the official documentation.